// ui/src/app/model/phone.ts
export class Phone {
    id: string;
    name: string;
    type: string;
    vendor: string;
    macAddress: string;
    ipAddress: string;
    protocol: string;
    model: string;
    subModel: string;
    firmwareVersion: string;
    location: string;
    group: string;
    linesList: any = [];
    username: string;
    password: string;
    ipSixAddress: string;
    primaryUser: string;
    primaryExtension: string;
    primaryDid: string;
    outBoundCallNumber: string;
    onPOINTOprId: string = null;
    oprDto: any;
    callServerManaged: boolean;
    callServerUnManged: boolean;
    callServerPartiallyManaged: boolean;
    remoteCaptureUsername: string;
    remoteCapturePassword: string;
    remoteCapturePort: string;
    remoteCaptureInterface: string;
    uri: string;
    freezeMAC: string;
    firmwareUpgradeStatus: string;
    phoneState: string;
}